Update app imprint from Sharezone GbR to Sharezone UG
