--- app/lib/settings/src/subpages/imprint/gateway/imprint_gateway.dart
@@ -1,22 +1,0 @@
-// Copyright (c) 2022 Sharezone UG (haftungsbeschränkt)
-// Licensed under the EUPL-1.2-or-later.
-//
-// You may obtain a copy of the Licence at:
-// https://joinup.ec.europa.eu/software/page/eupl
-//
-// SPDX-License-Identifier: EUPL-1.2
-
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:sharezone/settings/src/subpages/imprint/models/imprint.dart';
-
-class ImprintGateway {
-  final FirebaseFirestore _firestore;
-
-  ImprintGateway(this._firestore);
-
-  Stream<Imprint> get imprintStream => _firestore
-      .collection('Legal')
-      .doc('imprint')
-      .snapshots()
-      .map((doc) => Imprint.fromDocumentSnapshot(doc));
-}
