+++ app/lib/settings/src/subpages/imprint/page/imprint_page.dart
@@ -6,52 +6,28 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
+import 'package:legal/legal.dart';
 import 'package:sharezone/groups/src/widgets/contact_support.dart';
-import 'package:sharezone/settings/src/subpages/imprint/bloc/imprint_bloc.dart';
-import 'package:sharezone/settings/src/subpages/imprint/bloc/imprint_bloc_factory.dart';
-import 'package:sharezone/settings/src/subpages/imprint/models/imprint.dart';
 import 'package:sharezone_utils/launch_link.dart';
 
-class ImprintPage extends StatefulWidget {
+class ImprintPage extends StatelessWidget {
   static const tag = 'imprint-page';
 
   const ImprintPage({super.key});
 
   @override
-  State createState() => _ImprintPageState();
-}
-
-class _ImprintPageState extends State<ImprintPage> {
-  late ImprintBloc bloc;
-
-  @override
-  void initState() {
-    super.initState();
-    bloc = BlocProvider.of<ImprintBlocFactory>(context).create();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    final offlineData = Imprint.offline().asMarkdown;
     return Scaffold(
       appBar: AppBar(title: const Text("Impressum")),
       body: SingleChildScrollView(
         padding: const EdgeInsets.all(12),
         child: SafeArea(
-          child: StreamBuilder<String>(
-            initialData: offlineData,
-            stream: bloc.markdownStream,
-            builder: (context, snapshot) {
-              final data = snapshot.data ?? offlineData;
-              return MarkdownBody(
-                data: data,
-                onTapLink: (link, _, __) => launchURL(link),
-                selectable: true,
-              );
-            },
+          child: MarkdownBody(
+            data: markdownImprint,
+            onTapLink: (text, href, __) => launchURL(href!),
+            selectable: true,
           ),
         ),
       ),
