+++ app/pubspec.yaml
@@ -128,6 +128,8 @@ dependencies:
     path: ../lib/key_value_store
   last_online_reporting:
     path: ../lib/last_online_reporting
+  legal:
+    path: ../lib/legal
   logging: ^1.1.1
   # Can be replaced once Flutter officially supports the new Material Symbols,
   # see https://github.com/flutter/flutter/issues/102560.
