+++ lib/legal/lib/src/imprint/imprint.dart
@@ -1,0 +1,29 @@
+// Copyright (c) 2023 Sharezone UG (haftungsbeschränkt)
+// Licensed under the EUPL-1.2-or-later.
+//
+// You may obtain a copy of the Licence at:
+// https://joinup.ec.europa.eu/software/page/eupl
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+const markdownImprint = """### Adresse
+
+Sharezone UG (haftungsbeschränkt)
+Speditionstr. 15A
+40221 Düsseldorf
+
+### Geschäftsführer and redaktionell Verantwortlicher:
+Nils Reichardt & Jonas Sander
+
+### Handelsregister
+Registergericht: Amtsgericht Düsseldorf
+Registernummer: HRB 91086
+
+### Kontakt:
+Mobil: +49 1522 9504121
+E-Mail: support@sharezone.net
+
+### Umsatzsteuer-ID:
+
+Umsatzsteuer-Identifikationsnummer gemäß §27a Umsatzsteuergesetz: DE338063943
+""";
